BookList and Book classes are separated in two modules
