+++ index.js
@@ -1,6 +1,6 @@
 // class for Awesome books
 import { showAddbook, showContact, showList } from './modules/navBar_controls.js';
-import { BookList } from './modules/classes.js';
+import BookList from './modules/BookListClass.js';
 
 const myBookList = new BookList();
 myBookList.initialize();
